+++ checker-qual/src/main/java/org/checkerframework/checker/signature/qual/CanonicalName.java
@@ -41,9 +41,5 @@ import org.checkerframework.framework.qual.SubtypeOf;
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
-@SubtypeOf({
-  FullyQualifiedName.class,
-  CanonicalNameOrEmpty.class,
-  CanonicalNameOrPrimitiveType.class
-})
+@SubtypeOf({CanonicalNameOrEmpty.class})
 public @interface CanonicalName {}