+++ checker/tests/signature/ClassGetNameBinaryName.java
@@ -1,4 +1,3 @@
-import org.checkerframework.checker.signature.qual.CanonicalNameAndBinaryName;
 import org.checkerframework.checker.signature.qual.DotSeparatedIdentifiers;
 import org.checkerframework.checker.signature.qual.PrimitiveType;
 
@@ -65,37 +64,37 @@ public class ClassGetNameBinaryName {
 
   class TestGetCanonicalName {
 
-    @CanonicalNameAndBinaryName String s1 = ClassGetNameBinaryName.class.getCanonicalName();
+    @DotSeparatedIdentifiers String s1 = ClassGetNameBinaryName.class.getCanonicalName();
 
-    @CanonicalNameAndBinaryName String s2a = Integer.class.getCanonicalName();
+    @DotSeparatedIdentifiers String s2a = Integer.class.getCanonicalName();
 
-    @CanonicalNameAndBinaryName String s2b = java.lang.Integer.class.getCanonicalName();
+    @DotSeparatedIdentifiers String s2b = java.lang.Integer.class.getCanonicalName();
 
-    @CanonicalNameAndBinaryName String s4a = Boolean.class.getCanonicalName();
+    @DotSeparatedIdentifiers String s4a = Boolean.class.getCanonicalName();
 
     // :: error: (assignment)
     @PrimitiveType String s4b = Boolean.class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String s12 = Nested.class.getCanonicalName();
+    @DotSeparatedIdentifiers String s12 = Nested.class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String s13 = Inner.class.getName();
+    @DotSeparatedIdentifiers String s13 = Inner.class.getName();
 
     // Primitive types
 
     @PrimitiveType String prim1 = int.class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String prim2 = int.class.getCanonicalName();
+    @DotSeparatedIdentifiers String prim2 = int.class.getCanonicalName();
 
     @PrimitiveType String prim3 = boolean.class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String prim4 = boolean.class.getCanonicalName();
+    @DotSeparatedIdentifiers String prim4 = boolean.class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String prim5 = void.class.getCanonicalName();
+    @DotSeparatedIdentifiers String prim5 = void.class.getCanonicalName();
 
     // :: error: (assignment)
     @PrimitiveType String prim6 = void.class.getCanonicalName();
@@ -103,18 +102,18 @@ public class ClassGetNameBinaryName {
     // Arrays
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String s6 = int[].class.getCanonicalName();
+    @DotSeparatedIdentifiers String s6 = int[].class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String s7 = int[][].class.getCanonicalName();
+    @DotSeparatedIdentifiers String s7 = int[][].class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String s8 = boolean[].class.getCanonicalName();
+    @DotSeparatedIdentifiers String s8 = boolean[].class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String s9 = Integer[].class.getCanonicalName();
+    @DotSeparatedIdentifiers String s9 = Integer[].class.getCanonicalName();
 
     // :: error: (assignment)
-    @CanonicalNameAndBinaryName String s10 = Boolean[].class.getCanonicalName();
+    @DotSeparatedIdentifiers String s10 = Boolean[].class.getCanonicalName();
   }
 }